date modifications and move data bug fix

+++ s3_move_data.py
@@ -11,6 +11,11 @@ bucket_name = dbutils.widgets.get("bucket")
 source_folder = dbutils.widgets.get("source_folder")
 destination_folder = dbutils.widgets.get("destination_folder")
 
+print(region)
+print(bucket_name)
+print(source_folder)
+print(destination_folder)
+
 # COMMAND ----------
 
 #Creating Session With Boto3.
@@ -27,6 +32,7 @@ bucket = s3.Bucket(bucket_name)
 
 for obj in bucket.objects.filter(Prefix=source_folder):
     source_key = obj.key
-    destination_key = source_key.replace(source_folder, destination_folder, 1)
-    s3.Object(bucket_name, destination_key).copy_from(CopySource={'Bucket': bucket_name, 'Key': source_key})
-    s3.Object(bucket_name, source_key).delete()
+    destination_key = source_key.replace(source_folder, destination_folder, 1)    
+    if source_key != source_folder:
+        s3.Object(bucket_name, destination_key).copy_from(CopySource={'Bucket': bucket_name, 'Key': source_key})
+        s3.Object(bucket_name, source_key).delete()
